+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+* Toggle iOS FWDebug when double clicking the inspector if supported.
+* Added registerUrl method to register url with optional callback.
+
 ## 1.0.2
 
 * Keep the inspector position and do not reset it.